backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
boto3==1.34.162
botocore==1.34.162
requests==2.32.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1